Experimental support for specific backgrounds on specific default pages
In my case for now I added a check if the defualt card opened has wallbox in it's name to show a less Anime filled list of fitting backgrounds for the person that wants to check my monthly kWh I charged into my e-car. Works nicely it combined with a automation that sets a default theme.